tse_price_fetcher/requirements.txt
openpyxl>=3.1.0
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pandas>=2.1.0
python-dateutil>=2.8.0
tvdatafeed>=2.1.0
requests>=2.31.0
